Added splunk::purge
Purge will get rid of everything SPLUNK on your system. If you have an
install and want to get rid if it, and any splunk universal forwarder, use purge. It
also deletes everything connected to the Splunk RPM, so if you've built your own custom
RPMs to deploy TA's, it purges the heck out of them