Stop managing etc/auth dir contents
The current release installs a directory full of expired certificates into
etc/auth. This breaks the kvstore on the search/indexer machines, because
mongodb won't start with an expired cert.
Additionally, splunk seems to regenerate at least some of these certificates
when strting up, leading puppet to fight splunk over the contents and
frequent service restarts.

I can't see a good reason for this module to be managing the contents of the
certificate files; a job best left to the profile or glue module that uses
this splunk module, therefore I've removed it entirely.